Change all error messages to be in a class in lib directory
