+++ app/controllers/api/v1/items/search_controller.rb
@@ -5,7 +5,7 @@ class Api::V1::Items::SearchController < ApplicationController
   def index
     if search_by_name_only?
       item = Item.search_results(params[:name])
-      return json_response(ItemSerializer.items_error_message(params[:name]), 400) if item.nil?
+      return json_response(ErrorMessage.items_error_message(params[:name]), 400) if item.nil?
 
       item = Item.search_results(params[:name])
       json_response(ItemSerializer.new(item))
@@ -13,10 +13,10 @@ class Api::V1::Items::SearchController < ApplicationController
     else
 
       if searching_for_name_and_price_range?
-        json_response(ItemSerializer.item_min_price_search_error, :bad_request)
+        json_response(ErrorMessage.item_min_price_search_error, :bad_request)
 
       elsif min_price_out_of_bounds?
-        json_response(ItemSerializer.item_min_price_too_big)
+        json_response(ErrorMessage.item_min_price_too_big)
 
       elsif valid_min_and_max_price?
         item = Item.find_by_min_max_price(params[:min_price].to_i, params[:max_price].to_i)
@@ -24,13 +24,13 @@ class Api::V1::Items::SearchController < ApplicationController
         json_response(ItemSerializer.new(item))
 
       elsif missing_max_price?
-        return json_response(ItemSerializer.item_min_price_search_error, :bad_request) if params[:min_price].to_i <= 0
+        return json_response(ErrorMessage.item_min_price_search_error, :bad_request) if params[:min_price].to_i <= 0
 
         item = Item.find_by_min_price(params[:min_price].to_i)
         json_response(ItemSerializer.new(item))
 
       elsif missing_min_price?
-        return json_response(ItemSerializer.item_min_price_search_error, :bad_request) if params[:max_price].to_i <= 0
+        return json_response(ErrorMessage.item_min_price_search_error, :bad_request) if params[:max_price].to_i <= 0
 
         item = Item.find_by_max_price(params[:max_price].to_i)
         json_response(ItemSerializer.new(item))
