+++ app/controllers/api/v1/merchant_items_controller.rb
@@ -1,7 +1,7 @@
 class Api::V1::MerchantItemsController < ApplicationController
   def index
     if merchant_id_a_string?
-      json_response(MerchantSerializer.merchant_id_string_error('id', params[:merchant_id]), :not_found)
+      json_response(ErrorMessage.merchant_id_string_error('id', params[:merchant_id]), :not_found)
     else
       merchant = Merchant.find(params[:merchant_id])
       json_response(ItemSerializer.new(merchant.items))
