+++ app/controllers/api/v1/revenue/unshipped_controller.rb
@@ -1,9 +1,9 @@
 class Api::V1::Revenue::UnshippedController < ApplicationController
   def index
     if quantity_missing?
-      json_response(UnshippedRevenueSerializer.unshipped_invoices_error, :bad_request)
+      json_response(ErrorMessage.unshipped_invoices_error, :bad_request)
     elsif quantity_is_string?
-      json_response(UnshippedRevenueSerializer.unshipped_invoices_error, :bad_request)
+      json_response(ErrorMessage.unshipped_invoices_error, :bad_request)
     else
       invoices = Invoice.unshipped_revenue
       json_response(UnshippedRevenueSerializer.new(invoices))
