+++ app/serializers/unshipped_revenue_serializer.rb
@@ -4,12 +4,4 @@ class UnshippedRevenueSerializer
   attributes :potential_revenue do |invoice|
     invoice.revenue
   end
-
-  def self.unshipped_invoices_error
-    { "data": {
-      "message": "request could not be completed",
-      "code": 400,
-      "errors": [
-        "wrong data type does not exist"]}}
-  end
 end
